requirements.txt
flask
pandas
scikit-learn
joblib
numpy
python-dotenv